renamed spark_wiring_progmen to spark_wiring_arduino. Thanks @technobly!
