+++ wiring/inc/spark_wiring_arduino.h
